Suppress irrelevant gcc warning (mpls).
Modern gcc can warn about MPLS array accessed above its defined size
(which is MPLS_DEPTH) in Elements switch:

  warning: array subscript is above array bounds [-Warray-bounds]

Disabling this warning is safe, because, these statements only called
for a valid array length (due to template definition is based on
MPLS_DEPTH), so, supposedly over-reading code is never called.
For a double safety, sanity checking is added for --promisc-mpls value
(which defines MPLS_DEPTH) to be in range of 1..10.