Use strscpy rather than strlcpy (keep as fallback).
strscpy has been around since 4.3, and strlcpy got removed in 6.8.

This patch switches to the plainly preferred strscpy whenever it is
available, but will define that to strlcpy if strscpy (sized_strscpy due
to strscpy really being a macro) is not available.

Ideally we want to compile a test-call, but kbuild_test_symbol() doesn't
work that way.

We could also simply go "#ifndef strscpy ... # define strscpy strlcpy"
if that would be preferred.

Signed-off-by: Jaco Kroon <jaco@uls.co.za>